backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
httpx==0.26.0
websockets==12.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-genai==0.3.0
pillow==10.2.0
